Collapse all the code inside executor_traits.hpp
